Allow duplicated uploads #19
This change allows an artifact n to be contained in several bundles. n will not be deleted until the last bundle it belongs to is deleted. In addition a lone upload will cause a keep_alive flag to be set to make sure it's not deleted along with a bundle delete.